+++ CHANGELOG.md
@@ -4,11 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] / 2023-02-27 - 2023-02-28
+## [1.0.0] / 2023-02-27 - 2023-03-24
 ### Features
-- Application
-- Console
-- Shared
+### Application
+[x] PipeServer - Shared
+[x] Run Tests - ricaun.NUnit
+[] Filter Tests
+[x] Copy Zip Folder
+[] Copy back Zip Folder
+### Console
+[x] PipeClient - Shared
+[x] RevitInstallation - Get Revit Installed
+[x] CommandLine - Run File Tests
+[x] CommandLine - Get File Tests FullName
+[x] CommandLine - Output to Console/File
+[x] CommandLine - Logger enable/disable
+[x] CommandLine - Select RevitVersion
+[x] CommandLine - Force Open/Close/Wait
+[] CommandLine - Run Tests with Filter
+### Shared
+[x] PipeServer/Client
+[] Fody.PropertyChanged
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0